Moving domain objects into a dedicated package
