+++ src/main/java/com/github/vkorobkov/jfixtures/domain/Table.java
@@ -1,4 +1,4 @@
-package com.github.vkorobkov.jfixtures.loader;
+package com.github.vkorobkov.jfixtures.domain;
 
 import lombok.Getter;
 import lombok.val;
