+++ src/main/java/com/github/vkorobkov/jfixtures/domain/ValueType.java
@@ -1,4 +1,4 @@
-package com.github.vkorobkov.jfixtures.loader;
+package com.github.vkorobkov.jfixtures.domain;
 
 public enum ValueType {
     AUTO,
