+++ src/main/java/com/github/vkorobkov/jfixtures/loader/MapRowsLoader.java
@@ -1,5 +1,7 @@
 package com.github.vkorobkov.jfixtures.loader;
 
+import com.github.vkorobkov.jfixtures.domain.Row;
+import com.github.vkorobkov.jfixtures.domain.Value;
 import com.github.vkorobkov.jfixtures.util.StreamUtil;
 import lombok.AllArgsConstructor;
