+++ src/main/java/com/github/vkorobkov/jfixtures/sql/SqlBase.java
@@ -1,11 +1,11 @@
 package com.github.vkorobkov.jfixtures.sql;
 
 import com.github.vkorobkov.jfixtures.config.structure.tables.CleanMethod;
+import com.github.vkorobkov.jfixtures.domain.Value;
+import com.github.vkorobkov.jfixtures.domain.ValueType;
 import com.github.vkorobkov.jfixtures.instructions.CleanTable;
 import com.github.vkorobkov.jfixtures.instructions.CustomSql;
 import com.github.vkorobkov.jfixtures.instructions.InsertRow;
-import com.github.vkorobkov.jfixtures.loader.Value;
-import com.github.vkorobkov.jfixtures.loader.ValueType;
 import com.github.vkorobkov.jfixtures.util.SqlUtil;
 import lombok.SneakyThrows;
 import lombok.val;
