+++ src/test/groovy/com/github/vkorobkov/jfixtures/instructions/InsertRowTest.groovy
@@ -1,6 +1,6 @@
 package com.github.vkorobkov.jfixtures.instructions
 
-import com.github.vkorobkov.jfixtures.loader.Value
+import com.github.vkorobkov.jfixtures.domain.Value
 import spock.lang.Specification
 
 class InsertRowTest extends Specification {
