+++ src/test/groovy/com/github/vkorobkov/jfixtures/loader/RowTest.groovy
@@ -1,5 +1,7 @@
 package com.github.vkorobkov.jfixtures.loader
 
+import com.github.vkorobkov.jfixtures.domain.Row
+import com.github.vkorobkov.jfixtures.domain.Value
 import nl.jqno.equalsverifier.EqualsVerifier
 import spock.lang.Specification
