+++ src/test/groovy/com/github/vkorobkov/jfixtures/loader/TableTest.groovy
@@ -1,5 +1,7 @@
 package com.github.vkorobkov.jfixtures.loader
 
+import com.github.vkorobkov.jfixtures.domain.Row
+import com.github.vkorobkov.jfixtures.domain.Table
 import spock.lang.Specification
 
 class TableTest extends Specification {
