+++ src/test/groovy/com/github/vkorobkov/jfixtures/sql/dialects/MicrosoftSqlTest.groovy
@@ -1,14 +1,13 @@
 package com.github.vkorobkov.jfixtures.sql.dialects
 
 import com.github.vkorobkov.jfixtures.config.structure.tables.CleanMethod
+import com.github.vkorobkov.jfixtures.domain.Value
 import com.github.vkorobkov.jfixtures.instructions.CleanTable
 import com.github.vkorobkov.jfixtures.instructions.InsertRow
-import com.github.vkorobkov.jfixtures.loader.Value
 import com.github.vkorobkov.jfixtures.sql.Appender
 import com.github.vkorobkov.jfixtures.sql.Sql
 import com.github.vkorobkov.jfixtures.sql.appenders.StringAppender
 import spock.lang.Specification
-
 
 class MicrosoftSqlTest extends Specification {
     Sql sql
