+++ src/test/groovy/com/github/vkorobkov/jfixtures/sql/dialects/Sql99Test.groovy
@@ -1,9 +1,9 @@
 package com.github.vkorobkov.jfixtures.sql.dialects
 
 import com.github.vkorobkov.jfixtures.config.structure.tables.CleanMethod
+import com.github.vkorobkov.jfixtures.domain.Value
 import com.github.vkorobkov.jfixtures.instructions.CleanTable
 import com.github.vkorobkov.jfixtures.instructions.InsertRow
-import com.github.vkorobkov.jfixtures.loader.Value
 import com.github.vkorobkov.jfixtures.sql.Appender
 import com.github.vkorobkov.jfixtures.sql.Sql
 import com.github.vkorobkov.jfixtures.sql.appenders.StringAppender
